+++ medusa/helpers/__init__.py
@@ -46,7 +46,7 @@ from medusa.helper.common import (episode_num, http_code_description, media_exte
                                   pretty_file_size, subtitle_extensions)
 from medusa.helpers.utils import generate
 from medusa.imdb import Imdb
-from medusa.indexers.indexer_exceptions import IndexerException
+from medusa.indexers.exceptions import IndexerException
 from medusa.logger.adapters.style import BraceAdapter, BraceMessage
 from medusa.session.core import MedusaSafeSession
 from medusa.show.show import Show
@@ -205,7 +205,7 @@ def search_indexer_for_show_id(show_name, indexer=None, series_id=None, ui=None)
     :param ui: Custom UI for indexer use
     :return:
     """
-    from medusa.indexers.indexer_api import indexerApi
+    from medusa.indexers.api import indexerApi
     show_names = [re.sub('[. -]', ' ', show_name)]
 
     # Query Indexers for each search term and build the list of results
@@ -1075,8 +1075,8 @@ def real_path(path):
 
 def validate_show(show, season=None, episode=None):
     """Reindex show from originating indexer, and return indexer information for the passed episode."""
-    from medusa.indexers.indexer_api import indexerApi
-    from medusa.indexers.indexer_exceptions import IndexerEpisodeNotFound, IndexerSeasonNotFound, IndexerShowNotFound
+    from medusa.indexers.api import indexerApi
+    from medusa.indexers.exceptions import IndexerEpisodeNotFound, IndexerSeasonNotFound, IndexerShowNotFound
     indexer_lang = show.lang
 
     try:
@@ -1538,7 +1538,7 @@ def get_tvdb_from_id(indexer_id, indexer):
 
 
 def get_showname_from_indexer(indexer, indexer_id, lang='en'):
-    from medusa.indexers.indexer_api import indexerApi
+    from medusa.indexers.api import indexerApi
     indexer_api_params = indexerApi(indexer).api_params.copy()
     if lang:
         indexer_api_params['language'] = lang
