+++ medusa/indexers/tvmaze/exceptions.py
