+++ medusa/metadata/mede8er.py
@@ -10,8 +10,8 @@ import os
 from medusa import helpers
 from medusa.helper.common import dateFormat, episode_num, replace_extension
 from medusa.helper.exceptions import ex
-from medusa.indexers.indexer_api import indexerApi
-from medusa.indexers.indexer_exceptions import IndexerEpisodeNotFound, IndexerSeasonNotFound
+from medusa.indexers.api import indexerApi
+from medusa.indexers.exceptions import IndexerEpisodeNotFound, IndexerSeasonNotFound
 from medusa.logger.adapters.style import BraceAdapter
 from medusa.metadata import media_browser
