+++ medusa/scene_exceptions.py
@@ -13,8 +13,8 @@ import adba
 
 from medusa import app, db
 from medusa.helpers import sanitize_scene_name
-from medusa.indexers.indexer_api import indexerApi
-from medusa.indexers.indexer_config import INDEXER_TVDBV2
+from medusa.indexers.api import indexerApi
+from medusa.indexers.config import INDEXER_TVDBV2
 from medusa.logger.adapters.style import BraceAdapter
 from medusa.session.core import MedusaSafeSession
