+++ medusa/tv/base.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 import threading
 from builtins import object
 
-from medusa.indexers.indexer_config import INDEXER_TVDBV2
+from medusa.indexers.config import INDEXER_TVDBV2
 
 
 class Identifier(object):
