+++ tests/apiv2/test_config.py
@@ -8,7 +8,7 @@ import pkgutil
 import sys
 
 from medusa import app, classes, common, db, helpers, logger, metadata
-from medusa.indexers.indexer_config import INDEXER_TVDBV2
+from medusa.indexers.config import INDEXER_TVDBV2
 from medusa.common import cpu_presets
 from medusa.helpers.utils import int_default
 from medusa.sbdatetime import date_presets, time_presets
